* migrated the major part of the IMR CLI from imr/cmd/ to tdns/cli/
