+++ agent/README.md
@@ -1,32 +1,59 @@
 # TDNS-AGENT
 
-**TDNS-AGENT** is a slightly more limited version of **TDNS-SERVER**.
+**tdns-agent** is a slightly more limited version of **tdns-auth**.
 Primarily the limitation is that it is only able to operate as a
 secondary, not a primary. The intent is to enable zones that have a
 well-working zone production pipeline to easily attach the
-*tdns-agent* as an additional secondary and thereby gain access to the
+**tdns-agent** as an additional secondary and thereby gain access to the
 TDNS features without the need to modify the zone production.
 
-Design constraints for the TDNS Agent:
+## Use Cases
+
+1. Update the delegation information in the parent (for signed and unsigned zones).
+
+   Detect changes to delegation information in a zone (i.e. changes to the NS RRset
+   or possibly glue) and initiate action to update the corresponding information in
+   the parent zone. To do this, **tdns-agent** looks for a DSYNC RRset for the child
+   in the parent zone (i.e. child._dsync.parent. DSYNC) and examines the available
+   synhronization services (typically NOTIFY and/or UPDATE).
+
+   Then **tdns-agent** will either create and send a NOTIFY(CSYNC) to the parent
+   NOTIFY receiver, or create, sign and send a DNS UPDATE to the parent UPDATE receiver.
+
+2. Update the DS record in the parent (for a signed zone).
+
+   Same thing, but for CDS records. I.e. **tdns-agent** should detect publication of a new
+   child.parent. CDS record and then either create and send a NOTIFY(CDS) to the parent
+   NOTIFY receiver or create, sign and send a DNS UPDATE to the parent UPDATE receiver. [NYI]
+
+3. Multi-provider synchronization.
+
+   **tdns-agent** is able to watch the HSYNC RRset in the zone and take action on changes.
+   It will use the data in the HSYNC RRset to identify and establish secure communication
+   with remote agents. Once communication is established the agents are able to share
+   information between themselves to enable synchronization of the NS, DNSKEY, CDS and CSYNC
+   RRsets.
+
+   Note that as **tdns-agent** is prohibited from modifying any zone any changes that result
+   from synchronization with other agents will be submitted (via an API) to **tdns-combiner**
+   that does the actual modifcations to the zone.
+
+## Design Constraints for **tdns-agent**:
 
 1. Essentially all configuration errors should be fatal. There is no
    point to an agent that is running on a partially broken config.
 
-2. The **tdns-agent** can only serve secondary zones. Any primary zone in the
+2. **tdns-agent** can only serve secondary zones. Any primary zone in the
    configuration should cause the agent to terminate.
 
-3. Can not make modifications to zones. I.e. the options
-   online-signing, publish-key, allow-updates and allow-child-updates are errors and
+3. Can not make modifications to zones. I.e. options like online-signing,
+   publish-key, allow-updates and allow-child-updates are errors and
    should cause the agent to terminate.
 
-4. What the agent CAN do is to detect changes to delegation
-   information and when that happens take action. Depending on what
-   schemes the parent supports the action is either to send a generalized
-   NOTIFY or create, sign and send a DNS UPDATE.
-
-5. Not having either of the options delegation-sync-parent and/or
-   delegation-sync-child for a zone is an error, as then there is no
-   point to the agent being configured to deal with that zone.
+4. What the agent CAN do is to detect changes to delegation information and
+   when that happens take action. Depending on what schemes the parent supports
+   the action is either to send a generalized NOTIFY (to the parent notify
+   receiver) or create, sign and send a DNS UPDATE (to the parent update receiver).
 
 ```mermaid
 graph TD
